requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-client==0.19.0
python-dotenv==1.0.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
aiohttp==3.8.5